chore: 型をinput_port, output_portに分離させる
