+++ apps/backend/internal/repository/store_repository.go
@@ -4,24 +4,16 @@ import (
 	"context"
 
 	"github.com/TeamH04/team-production/apps/backend/internal/domain"
+	"github.com/TeamH04/team-production/apps/backend/internal/usecase/output_port"
 	"gorm.io/gorm"
 )
-
-// StoreRepository はストアのデータアクセスを抽象化するインターフェース
-type StoreRepository interface {
-	FindAll(ctx context.Context) ([]domain.Store, error)
-	FindByID(ctx context.Context, id int64) (*domain.Store, error)
-	Create(ctx context.Context, store *domain.Store) error
-	Update(ctx context.Context, store *domain.Store) error
-	Delete(ctx context.Context, id int64) error
-}
 
 type storeRepository struct {
 	db *gorm.DB
 }
 
 // NewStoreRepository は StoreRepository の実装を生成します
-func NewStoreRepository(db *gorm.DB) StoreRepository {
+func NewStoreRepository(db *gorm.DB) output_port.StoreRepository {
 	return &storeRepository{db: db}
 }
